refactor(frontend): consolidate feeds modules into feedsService
Merges feed creation api, error normalization, utility functions, and preview loading/retries from four separate files under frontend/src/feeds/ into a unified feedsService.ts file.
