+++ frontend/src/__tests__/previewHydration.test.ts
@@ -4,7 +4,7 @@ import {
   loadPreviewItems,
   loadPreviewItemsWithRetry,
   normalizePreviewItems,
-} from '../feeds/previewHydration';
+} from '../feeds/feedsService';
 
 const feed = {
   id: 'feed-1',
